Major feature update: UI enhancements and modeling features
- Added professional icon system with SVG icons
- Implemented layout templates (Inventor, SolidWorks, CATIA styles)
- Enhanced dock system with save/load functionality
- Extended status bar with coordinates, units, snap mode
- Added enhanced tooltips with keyboard shortcuts
- Implemented 8 new modeling features (Sweep, Helix, Shell, Draft, Chamfer, Mirror, Thread, Rib)
- Updated CHANGELOG.md
- Code cleanup and documentation

+++ src/modules/mbd/MbdService.cpp
@@ -214,9 +214,9 @@ void MbdService::updateMbdVisibility(const std::string& part_id, bool show_annot
         show_annotations,
         show_datums,
         show_tolerances,
-        result.visible_annotations.size(),
-        result.visible_datums.size(),
-        result.visible_tolerances.size()
+        static_cast<int>(result.visible_annotations.size()),
+        static_cast<int>(result.visible_datums.size()),
+        static_cast<int>(result.visible_tolerances.size())
     };
 }
